safer typing for stride: int -> size_t
first pass, may have missed a bunch.

even if picture dimension (width/height) is reasonable, 'stride' can be
arbitrary large. 'size_t' is the natural type here, and makes calculations
like 'ptr + y * stride' safe.

+++ src/dsp/dsp.h
@@ -495,9 +495,9 @@ extern void (*WebPDispatchAlphaToGreen)(const uint8_t* alpha, int alpha_stride,
 // Extract the alpha values from 32b values in argb[] and pack them into alpha[]
 // (this is the opposite of WebPDispatchAlpha).
 // Returns true if there's only trivial 0xff alpha values.
-extern int (*WebPExtractAlpha)(const uint8_t* argb, int argb_stride,
+extern int (*WebPExtractAlpha)(const uint8_t* argb, size_t argb_stride,
                                int width, int height,
-                               uint8_t* alpha, int alpha_stride);
+                               uint8_t* alpha, size_t alpha_stride);
 
 // Pre-Multiply operation transforms x into x * A / 255  (where x=Y,R,G or B).
 // Un-Multiply operation transforms x into x * 255 / A.
